Changed some test names to reflect the numbers of the ranges
In order to be more clear, i have decided to put the numerical
representation of the number ranges, because i see it more clear and
readable.

I don’t like to put numbers in method names, but i think that the test
methods should be clear and this will help.