Refactoring code to make it more testable
Right now Im having a problem, I need to test if my bingo check cards,
but as Im thinking of it, I am not able to test the 2 test cases that
it presents.

I need to know if a card that has called “Bingo!” its really truth or
the client it’s just joking at my game, so I am thinking that it’s
needed to make a strong refactor, and improve my code in order to
separate responsabilities and be able to make the code easier to test.

Right now we have a big test class that test all we have, but Im going
to separate some of the test cases in another test class (to test the
card for example), making the tests more focused.