Adapted tests for third row in the bingo generator test
Here i have had some problems.

Im a newbie in python, so my approach in the columns generation was
with a set(), but here in this test i have seen that the set data
structure doesn’t respect the inspection order, so i have changed the
approach with a list, and a refactor to make the test pass green again.

My other tests were working well, so nothing more to fix here, and all
is green